Convert several physics automated tests with Base test level file converted to prefab file (#5138)
* Convert some physics automated tests to use prefab system


* revert changes


* Fix file names
