+++ AutomatedTesting/Gem/PythonTests/Physics/tests/collider/Collider_BoxShapeEditing.py
@@ -4,8 +4,8 @@ For complete copyright and license terms please see the LICENSE at the root of t
 
 SPDX-License-Identifier: Apache-2.0 OR MIT
 
-Test case ID : C4982802
-Test Case Title : Verify that the shape capsule can be selected from drop downlist and the value for its height and radius can be set after that
+Test case ID : C4982801
+Test Case Title : Verify that the shape Box can be selected from drop downlist and the value for its dimensions in x,y,z can be set after that
 
 """
 
@@ -19,19 +19,19 @@ class Tests():
 # fmt: on
 
 
-def Collider_CapsuleShapeEditting():
+def Collider_BoxShapeEditing():
     """
     Summary:
      Adding PhysX Collider and Shape components to test entity, then attempting to modify the shape's dimensions
 
     Expected Behavior:
-     Capsule shape can be selected for the Shape Component and the value for Height and Radius can be changed
+     Box shape can be selected for the Shape Component and the value for X, Y, and Z dimensions can be changed
 
     Test Steps:
      1) Load the empty level
      2) Create the test entity
      3) Add PhysX Collider component to test entity
-     4) Change the PhysX Collider shape
+     4) Change the PhysX Collider shape and store the original dimensions
      5) Modify the dimensions
      6) Verify they have been changed
 
@@ -51,30 +51,29 @@ def Collider_CapsuleShapeEditting():
     # Open 3D Engine Imports
     import azlmbr.math as math
 
-    CAPSULE_SHAPETYPE_ENUM = 2
-    # Note from Editor Console: Height must exceed twice the radius in capsule configuration
-    SIZE_RADIUS = 4.0
-    SIZE_HEIGHT = SIZE_RADIUS * 2 + 1.0
+    BOX_SHAPETYPE_ENUM = 1
+    SET_SIZE = 2.5
     SIZE_TOLERANCE = 0.5
 
-    def change_dimension_value(component, component_property_path, value):
-        Report.info(f"Attempting to set value for {component_property_path} to {value}")
-        component.set_component_property_value(component_property_path, value)
-        returning_value = component.get_component_property_value(component_property_path)
-        Report.info(f"Value for {component_property_path} is currently {returning_value}")
-        return returning_value
+    def check_dimensions_changed(set_dimension_value, modified_dimensions, tolerance):
+        def compare_values(value_name, set_value, grabbed_value, tolerance):
+            within_tolerance = math.Math_IsClose(set_value, grabbed_value, tolerance)
+            if not within_tolerance:
+                assert (
+                    False
+                ), f"The modified value for {value_name} was not within the allowed tolerance\nExpected:{set_value}\nActual: {grabbed_value}"
+            return within_tolerance
 
-    def check_dimension_change(value_name, value_set, grabbed_value, tolerance):
-        within_tolerance = math.Math_IsClose(value_set, grabbed_value, tolerance)
-        if not within_tolerance:
-            assert (
-                False
-            ), f"The modified value for {value_name} was not within the allowed tolerance\nExpected:{value_set}\nActual: {grabbed_value}"
-        return within_tolerance
+        # Check for X, Y, & Z values
+        x_value = compare_values("x_value", set_dimension_value, modified_dimensions.x, tolerance)
+        y_value = compare_values("y_value", set_dimension_value, modified_dimensions.y, tolerance)
+        z_value = compare_values("z_value", set_dimension_value, modified_dimensions.z, tolerance)
+
+        return x_value and y_value and z_value
 
     helper.init_idle()
     # 1) Load the empty level
-    helper.open_level("Physics", "Base")
+    helper.open_level("", "Base")
 
     # 2) Create the test entity
     test_entity = Entity.create_editor_entity("Test Entity")
@@ -84,22 +83,23 @@ def Collider_CapsuleShapeEditting():
     test_component = test_entity.add_component("PhysX Collider")
     Report.result(Tests.collider_added, test_entity.has_component("PhysX Collider"))
 
-    # 4) Change the PhysX Collider shape
-    test_component.set_component_property_value("Shape Configuration|Shape", CAPSULE_SHAPETYPE_ENUM)
-    add_check = test_component.get_component_property_value("Shape Configuration|Shape") == CAPSULE_SHAPETYPE_ENUM
+    # 4) Change the PhysX Collider shape and store the original dimensions
+    test_component.set_component_property_value("Shape Configuration|Shape", BOX_SHAPETYPE_ENUM)
+    add_check = test_component.get_component_property_value("Shape Configuration|Shape") == BOX_SHAPETYPE_ENUM
     Report.result(Tests.collider_shape_changed, add_check)
 
     # 5) Modify the dimensions
-    mod_height = change_dimension_value(test_component, "Shape Configuration|Capsule|Height", SIZE_HEIGHT)
-    mod_radius = change_dimension_value(test_component, "Shape Configuration|Capsule|Radius", SIZE_RADIUS)
+    Report.info(f"Attempting to set XYZ values to {SET_SIZE}")
+    test_component.set_component_property_value(
+        "Shape Configuration|Box|Dimensions", math.Vector3(SET_SIZE, SET_SIZE, SET_SIZE)
+    )
+    mod_dimensions = test_component.get_component_property_value("Shape Configuration|Box|Dimensions")
 
     # 6) Verify they have been changed
-    resulting_check = check_dimension_change(
-        "Height", SIZE_HEIGHT, mod_height, SIZE_TOLERANCE
-    ) and check_dimension_change("Radius", SIZE_RADIUS, mod_radius, SIZE_TOLERANCE)
-    Report.result(Tests.shape_dimensions_changed, resulting_check)
+    dimensions_successfully_changed = check_dimensions_changed(SET_SIZE, mod_dimensions, SIZE_TOLERANCE)
+    Report.result(Tests.shape_dimensions_changed, dimensions_successfully_changed)
 
 
 if __name__ == "__main__":
     from editor_python_test_tools.utils import Report
-    Report.start_test(Collider_CapsuleShapeEditting)
+    Report.start_test(Collider_BoxShapeEditing)
